Refactor: PositionDto 승격 및 PlaceResponseDto 제거
