+++ src/main/java/com/jigumulmi/place/domain/Place.java
@@ -5,7 +5,7 @@ import com.jigumulmi.admin.place.dto.request.AdminCreatePlaceRequestDto;
 import com.jigumulmi.banner.domain.BannerPlaceMapping;
 import com.jigumulmi.common.Timestamped;
 import com.jigumulmi.member.domain.Member;
-import com.jigumulmi.place.dto.response.PlaceResponseDto.PositionDto;
+import com.jigumulmi.place.dto.PositionDto;
 import com.jigumulmi.place.vo.District;
 import com.jigumulmi.place.vo.Region;
 import jakarta.persistence.CascadeType;
