+++ src/main/java/com/jigumulmi/place/dto/PositionDto.java
@@ -1,0 +1,16 @@
+package com.jigumulmi.place.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PositionDto {
+
+    private Double latitude;
+    private Double longitude;
+}
